Create entity management functions & struct
EntityCategory -> a category of entities with a common name (e.g. "enemy") and a map of component names to a list of components for every entity (e.g. "Pos": 1->[0,0] 2 -> [-50, 0] 3 -> [0, 50])
createEntityCategory -> create an entity category by passing a map of components common for all new entities
addEntity -> add an entity with the specified components